+++ python/docs/api.rst
@@ -17,3 +17,21 @@ Memory Resources
    :members:
    :undoc-members:
    :show-inheritance:
+
+Memory Allocators
+-----------------
+
+.. automodule:: rmm.allocators.cupy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: rmm.allocators.numba
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: rmm.allocators.torch
+   :members:
+   :undoc-members:
+   :show-inheritance:
